Demo CS: mocking server by default, flag to use real server

+++ examples/Demo_5_CS_Client.py
@@ -5,7 +5,13 @@ Client/Server demo with Pyfhel
 This demo shows an example of Client-Server interaction, where the client sends
 an encrypted vector and the server answers with the weighted average based on his
 weights.
+
+To run with real Client/Server separation (using flask and Demo_5bis_CS_Server.py),
+change the flag below to True.
 """
+
+USE_REAL_SERVER: bool = False
+
 # %% 1. Setup Client
 import numpy as np
 from Pyfhel import Pyfhel, PyCtxt
@@ -39,28 +45,54 @@ print(f"[Client] sending {HE_client=} and {cx=}")
 # %% 2. Setup Server
 
 print(f"[Client] launching server (could be launched separately)...")
-import subprocess, os
-from pathlib import Path
-dir = Path(os.path.realpath("__file__")).parent
-process = subprocess.Popen(
-    ["python", str(dir / "Demo_5bis_CS_Server.py")],
-    stderr=subprocess.STDOUT,
-)
-import time
-time.sleep(6)       # Wait for server initialization
+if(USE_REAL_SERVER):
+    import subprocess, os
+    from pathlib import Path
+    dir = Path(os.path.realpath("__file__")).parent
+    process = subprocess.Popen(
+        ["python", str(dir / "Demo_5bis_CS_Server.py")],
+        stderr=subprocess.STDOUT,
+    )
+    import time
+    time.sleep(6)       # Wait for server initialization
+else:
+    print(f"[Server] started!...")
 print("[Client] server initialized...")
 
-# %% 3. Launch a request to the server and process the response
+# %% 3. Launch a request to the server
 #  We map the bytes into strings based on https://stackoverflow.com/a/27527728
-r = requests.post('http://127.0.0.1:5000/fhe_mse',
-    json={
-        'context': s_context.decode('cp437'),
-        'pk': s_public_key.decode('cp437'),
-        'rlk':s_relin_key.decode('cp437'),
-        'rtk':s_rotate_key.decode('cp437'),
-        'cx': s_cx.decode('cp437'),
-    }
-)
+if(USE_REAL_SERVER):
+    r = requests.post('http://127.0.0.1:5000/fhe_mse',
+        json={
+            'context': s_context.decode('cp437'),
+            'pk': s_public_key.decode('cp437'),
+            'rlk':s_relin_key.decode('cp437'),
+            'rtk':s_rotate_key.decode('cp437'),
+            'cx': s_cx.decode('cp437'),
+        }
+    )
+else: # Mocking server code (from Demo_5bis_CS_Server.py)
+    # Read all bytestrings
+    HE_server = Pyfhel()
+    HE_server.from_bytes_context(s_context)
+    HE_server.from_bytes_public_key(s_public_key)
+    HE_server.from_bytes_relin_key(s_relin_key)
+    HE_server.from_bytes_rotate_key(s_rotate_key)
+    cx = PyCtxt(pyfhel=HE_server, bytestring=s_cx)
+    print(f"[Server] received {HE_server=} and {cx=}")
+
+    # Encode weights in plaintext
+    w = np.array([0.5, -1.5,   4,  5])
+    ptxt_w = HE_server.encode(w)
+
+    # Compute weighted average
+    c_mean = (cx * ptxt_w)
+    c_mean /= 4  # 4
+    c_mean += (c_mean >> 1)   # cumulative sum
+    c_mean += (c_mean >> 2)   # element [3] contains the result
+    print(f"[Server] Average computed! Responding: {c_mean=}")
+
+# %% 4. Process Response
 # Decrypting result
 c_res = PyCtxt(pyfhel=HE_client, bytestring=r.text.encode('cp437'))
 res = HE_client.decryptFrac(c_res)
@@ -71,7 +103,8 @@ expected = np.mean(x*w)
 print(f"[Client] Response received! Result is {np.round(res[3], 4)}, should be {expected}")
 
 
-# %% 4. Stop server
-process.kill()
+# %% 5. Stop server
+if USE_REAL_SERVER:
+    process.kill()
 
 # sphinx_gallery_thumbnail_path = 'static/thumbnails/encrypting.jpg'
